+++ packages/keyring-controller/src/KeyringController.test.ts
@@ -191,17 +191,30 @@ describe('KeyringController', () => {
           );
         });
       });
-    });
 
-    it('should throw error with no HD keyring', async () => {
-      await withController(
-        { skipVaultCreation: true },
-        async ({ controller }) => {
+      it('should throw an error if there is no primary keyring', async () => {
+        await withController(async ({ controller, encryptor }) => {
+          await controller.setLocked();
+          jest
+            .spyOn(encryptor, 'decrypt')
+            .mockResolvedValueOnce([{ type: 'Unsupported', data: '' }]);
+          await controller.submitPassword('123');
+
           await expect(controller.addNewAccount()).rejects.toThrow(
             'No HD keyring found',
           );
-        },
-      );
+        });
+      });
+    });
+
+    it('should throw error when the controller is locked', async () => {
+      await withController(async ({ controller }) => {
+        await controller.setLocked();
+
+        await expect(controller.addNewAccount()).rejects.toThrow(
+          KeyringControllerError.ControllerLocked,
+        );
+      });
     });
 
     // Testing fix for bug #4157 {@link https://github.com/MetaMask/core/issues/4157}
@@ -353,6 +366,17 @@ describe('KeyringController', () => {
         });
       });
     });
+
+    it('should throw error when the controller is locked', async () => {
+      await withController(async ({ controller }) => {
+        const keyring = controller.getKeyringsByType(KeyringTypes.hd)[0];
+        await controller.setLocked();
+
+        await expect(
+          controller.addNewAccountForKeyring(keyring as EthKeyring<Json>),
+        ).rejects.toThrow(KeyringControllerError.ControllerLocked);
+      });
+    });
   });
 
   describe('addNewKeyring', () => {
@@ -374,6 +398,16 @@ describe('KeyringController', () => {
             'KeyringController - No keyringBuilder found for keyring. Keyring type: fake',
           );
         });
+      });
+    });
+
+    it('should throw error when the controller is locked', async () => {
+      await withController(async ({ controller }) => {
+        await controller.setLocked();
+
+        await expect(controller.addNewKeyring(KeyringTypes.hd)).rejects.toThrow(
+          KeyringControllerError.ControllerLocked,
+        );
       });
     });
   });
@@ -630,6 +664,16 @@ describe('KeyringController', () => {
         expect(listener.called).toBe(true);
       });
     });
+
+    it('should throw error when the controller is locked', async () => {
+      await withController(async ({ controller }) => {
+        await controller.setLocked();
+
+        await expect(controller.setLocked()).rejects.toThrow(
+          KeyringControllerError.ControllerLocked,
+        );
+      });
+    });
   });
 
   describe('exportSeedPhrase', () => {
@@ -670,6 +714,16 @@ describe('KeyringController', () => {
             );
           });
         });
+      });
+    });
+
+    it('should throw error when the controller is locked', async () => {
+      await withController(async ({ controller }) => {
+        await controller.setLocked();
+
+        await expect(controller.exportSeedPhrase(password)).rejects.toThrow(
+          KeyringControllerError.ControllerLocked,
+        );
       });
     });
   });
@@ -751,6 +805,16 @@ describe('KeyringController', () => {
         expect(accounts).toStrictEqual(initialAccount);
       });
     });
+
+    it('should throw error when the controller is locked', async () => {
+      await withController(async ({ controller }) => {
+        await controller.setLocked();
+
+        await expect(controller.getAccounts()).rejects.toThrow(
+          KeyringControllerError.ControllerLocked,
+        );
+      });
+    });
   });
 
   describe('getEncryptionPublicKey', () => {
@@ -790,6 +854,18 @@ describe('KeyringController', () => {
             );
           },
         );
+      });
+    });
+
+    it('should throw error when the controller is locked', async () => {
+      await withController(async ({ controller, initialState }) => {
+        await controller.setLocked();
+
+        await expect(
+          controller.getEncryptionPublicKey(
+            initialState.keyrings[0].accounts[0],
+          ),
+        ).rejects.toThrow(KeyringControllerError.ControllerLocked);
       });
     });
   });
@@ -868,6 +944,24 @@ describe('KeyringController', () => {
         );
       });
     });
+
+    it('should throw error when the controller is locked', async () => {
+      await withController(async ({ controller, initialState }) => {
+        await controller.setLocked();
+
+        await expect(
+          controller.decryptMessage({
+            from: initialState.keyrings[0].accounts[0],
+            data: {
+              version: '1.0',
+              nonce: '123456',
+              ephemPublicKey: '0xabcdef1234567890',
+              ciphertext: '0xabcdef1234567890',
+            },
+          }),
+        ).rejects.toThrow(KeyringControllerError.ControllerLocked);
+      });
+    });
   });
 
   describe('getKeyringForAccount', () => {
@@ -889,7 +983,7 @@ describe('KeyringController', () => {
     });
 
     describe('when non-existing account is provided', () => {
-      it('should throw error', async () => {
+      it('should throw error if no account matches the address', async () => {
         await withController(async ({ controller }) => {
           await expect(
             controller.getKeyringForAccount(
@@ -901,19 +995,44 @@ describe('KeyringController', () => {
         });
       });
 
-      it('should throw an error if there are no keyrings', async () => {
-        await withController(
-          { skipVaultCreation: true },
-          async ({ controller }) => {
-            await expect(
-              controller.getKeyringForAccount(
-                '0x51253087e6f8358b5f10c0a94315d69db3357859',
-              ),
-            ).rejects.toThrow(
-              'KeyringController - No keyring found. Error info: There are no keyrings',
-            );
-          },
-        );
+      it('should throw an error if there is no keyring', async () => {
+        await withController(async ({ controller, encryptor }) => {
+          await controller.setLocked();
+          jest
+            .spyOn(encryptor, 'decrypt')
+            .mockResolvedValueOnce([{ type: 'Unsupported', data: '' }]);
+          await controller.submitPassword('123');
+
+          await expect(
+            controller.getKeyringForAccount(
+              '0x0000000000000000000000000000000000000000',
+            ),
+          ).rejects.toThrow(
+            'KeyringController - No keyring found. Error info: There are no keyrings',
+          );
+        });
+      });
+
+      it('should throw an error if the controller is locked', async () => {
+        await withController(async ({ controller }) => {
+          await controller.setLocked();
+
+          await expect(
+            controller.getKeyringForAccount(
+              '0x51253087e6f8358b5f10c0a94315d69db3357859',
+            ),
+          ).rejects.toThrow(KeyringControllerError.ControllerLocked);
+        });
+      });
+    });
+
+    it('should throw error when the controller is locked', async () => {
+      await withController(async ({ controller, initialState }) => {
+        await controller.setLocked();
+
+        await expect(
+          controller.getKeyringForAccount(initialState.keyrings[0].accounts[0]),
+        ).rejects.toThrow(KeyringControllerError.ControllerLocked);
       });
     });
   });
@@ -942,6 +1061,16 @@ describe('KeyringController', () => {
         });
       });
     });
+
+    it('should throw error when the controller is locked', async () => {
+      await withController(async ({ controller }) => {
+        await controller.setLocked();
+
+        expect(() => controller.getKeyringsByType(KeyringTypes.hd)).toThrow(
+          KeyringControllerError.ControllerLocked,
+        );
+      });
+    });
   });
 
   describe('persistAllKeyrings', () => {
@@ -963,7 +1092,7 @@ describe('KeyringController', () => {
         await controller.setLocked();
 
         await expect(controller.persistAllKeyrings()).rejects.toThrow(
-          KeyringControllerError.MissingCredentials,
+          KeyringControllerError.ControllerLocked,
         );
       });
     });
@@ -1152,6 +1281,19 @@ describe('KeyringController', () => {
         });
       });
     });
+
+    it('should throw error when the controller is locked', async () => {
+      await withController(async ({ controller }) => {
+        await controller.setLocked();
+
+        await expect(
+          controller.importAccountWithStrategy(
+            AccountImportStrategy.privateKey,
+            [input, 'password'],
+          ),
+        ).rejects.toThrow(KeyringControllerError.ControllerLocked);
+      });
+    });
   });
 
   describe('removeAccount', () => {
@@ -1246,6 +1388,16 @@ describe('KeyringController', () => {
         );
       });
     });
+
+    it('should throw error when the controller is locked', async () => {
+      await withController(async ({ controller, initialState }) => {
+        await controller.setLocked();
+
+        await expect(
+          controller.removeAccount(initialState.keyrings[0].accounts[0]),
+        ).rejects.toThrow(KeyringControllerError.ControllerLocked);
+      });
+    });
   });
 
   describe('signMessage', () => {
@@ -1307,6 +1459,20 @@ describe('KeyringController', () => {
             );
           },
         );
+      });
+    });
+
+    it('should throw error when the controller is locked', async () => {
+      await withController(async ({ controller, initialState }) => {
+        await controller.setLocked();
+
+        await expect(
+          controller.signMessage({
+            from: initialState.keyrings[0].accounts[0],
+            data: '0x879a053d4800c6354e76c7985a865d2922c82fb5b3f4577b2fe08b998954f2e0',
+            origin: 'https://metamask.github.io',
+          }),
+        ).rejects.toThrow(KeyringControllerError.ControllerLocked);
       });
     });
   });
@@ -1377,6 +1543,20 @@ describe('KeyringController', () => {
             );
           },
         );
+      });
+    });
+
+    it('should throw error when the controller is locked', async () => {
+      await withController(async ({ controller, initialState }) => {
+        await controller.setLocked();
+
+        await expect(
+          controller.signPersonalMessage({
+            from: initialState.keyrings[0].accounts[0],
+            data: '0x879a053d4800c6354e76c7985a865d2922c82fb5b3f4577b2fe08b998954f2e0',
+            origin: 'https://metamask.github.io',
+          }),
+        ).rejects.toThrow(KeyringControllerError.ControllerLocked);
       });
     });
   });
@@ -1652,6 +1832,34 @@ describe('KeyringController', () => {
         );
       });
     });
+
+    it('should throw error when the controller is locked', async () => {
+      await withController(async ({ controller, initialState }) => {
+        await controller.setLocked();
+
+        await expect(
+          controller.signTypedMessage(
+            {
+              from: initialState.keyrings[0].accounts[0],
+              data: [
+                {
+                  type: 'string',
+                  name: 'Message',
+                  value: 'Hi, Alice!',
+                },
+                {
+                  type: 'uint32',
+                  name: 'A number',
+                  value: '1337',
+                },
+              ],
+              origin: 'https://metamask.github.io',
+            },
+            SignTypedDataVersion.V1,
+          ),
+        ).rejects.toThrow(KeyringControllerError.ControllerLocked);
+      });
+    });
   });
 
   describe('signTransaction', () => {
@@ -1740,6 +1948,19 @@ describe('KeyringController', () => {
         );
       });
     });
+
+    it('should throw error when the controller is locked', async () => {
+      await withController(async ({ controller, initialState }) => {
+        await controller.setLocked();
+
+        await expect(
+          controller.signTransaction(
+            buildMockTransaction(),
+            initialState.keyrings[0].accounts[0],
+          ),
+        ).rejects.toThrow(KeyringControllerError.ControllerLocked);
+      });
+    });
   });
 
   describe('prepareUserOperation', () => {
@@ -1816,6 +2037,20 @@ describe('KeyringController', () => {
             );
           },
         );
+      });
+    });
+
+    it('should throw error when the controller is locked', async () => {
+      await withController(async ({ controller, initialState }) => {
+        await controller.setLocked();
+
+        await expect(
+          controller.prepareUserOperation(
+            initialState.keyrings[0].accounts[0],
+            [],
+            executionContext,
+          ),
+        ).rejects.toThrow(KeyringControllerError.ControllerLocked);
       });
     });
   });
@@ -1905,6 +2140,32 @@ describe('KeyringController', () => {
         );
       });
     });
+
+    it('should throw error when the controller is locked', async () => {
+      await withController(async ({ controller, initialState }) => {
+        await controller.setLocked();
+
+        await expect(
+          controller.patchUserOperation(
+            initialState.keyrings[0].accounts[0],
+            {
+              sender: '0x4584d2B4905087A100420AFfCe1b2d73fC69B8E4',
+              nonce: '0x1',
+              initCode: '0x',
+              callData: '0x7064',
+              callGasLimit: '0x58a83',
+              verificationGasLimit: '0xe8c4',
+              preVerificationGas: '0xc57c',
+              maxFeePerGas: '0x87f0878c0',
+              maxPriorityFeePerGas: '0x1dcd6500',
+              paymasterAndData: '0x',
+              signature: '0x',
+            },
+            executionContext,
+          ),
+        ).rejects.toThrow(KeyringControllerError.ControllerLocked);
+      });
+    });
   });
 
   describe('signUserOperation', () => {
@@ -1989,6 +2250,32 @@ describe('KeyringController', () => {
         );
       });
     });
+
+    it('should throw error when the controller is locked', async () => {
+      await withController(async ({ controller, initialState }) => {
+        await controller.setLocked();
+
+        await expect(
+          controller.signUserOperation(
+            initialState.keyrings[0].accounts[0],
+            {
+              sender: '0x4584d2B4905087A100420AFfCe1b2d73fC69B8E4',
+              nonce: '0x1',
+              initCode: '0x',
+              callData: '0x7064',
+              callGasLimit: '0x58a83',
+              verificationGasLimit: '0xe8c4',
+              preVerificationGas: '0xc57c',
+              maxFeePerGas: '0x87f0878c0',
+              maxPriorityFeePerGas: '0x1dcd6500',
+              paymasterAndData: '0x',
+              signature: '0x',
+            },
+            executionContext,
+          ),
+        ).rejects.toThrow(KeyringControllerError.ControllerLocked);
+      });
+    });
   });
 
   describe('changePassword', () => {
@@ -2018,9 +2305,9 @@ describe('KeyringController', () => {
             async ({ controller }) => {
               await controller.setLocked();
 
-              await expect(controller.changePassword('')).rejects.toThrow(
-                KeyringControllerError.MissingCredentials,
-              );
+              await expect(async () =>
+                controller.changePassword(''),
+              ).rejects.toThrow(KeyringControllerError.ControllerLocked);
             },
           );
         });
@@ -2046,6 +2333,16 @@ describe('KeyringController', () => {
               ).rejects.toThrow(KeyringControllerError.WrongPasswordType);
             },
           );
+        });
+
+        it('should throw error when the controller is locked', async () => {
+          await withController(async ({ controller }) => {
+            await controller.setLocked();
+
+            await expect(async () =>
+              controller.changePassword('whatever'),
+            ).rejects.toThrow(KeyringControllerError.ControllerLocked);
+          });
         });
       }),
     );
@@ -2265,15 +2562,39 @@ describe('KeyringController', () => {
       });
     });
 
-    it('should throw error with no HD keyring', async () => {
+    it('should throw error if the controller is locked', async () => {
       await withController(
         { skipVaultCreation: true },
         async ({ controller }) => {
           await expect(controller.verifySeedPhrase()).rejects.toThrow(
-            'No HD keyring found',
+            KeyringControllerError.ControllerLocked,
           );
         },
       );
+    });
+
+    it('should throw an error if there is no primary keyring', async () => {
+      await withController(async ({ controller, encryptor }) => {
+        await controller.setLocked();
+        jest
+          .spyOn(encryptor, 'decrypt')
+          .mockResolvedValueOnce([{ type: 'Unsupported', data: '' }]);
+        await controller.submitPassword('123');
+
+        await expect(controller.verifySeedPhrase()).rejects.toThrow(
+          'No HD keyring found',
+        );
+      });
+    });
+
+    it('should throw error when the controller is locked', async () => {
+      await withController(async ({ controller }) => {
+        await controller.setLocked();
+
+        await expect(controller.verifySeedPhrase()).rejects.toThrow(
+          KeyringControllerError.ControllerLocked,
+        );
+      });
     });
   });
 
@@ -2457,6 +2778,18 @@ describe('KeyringController', () => {
         );
       });
     });
+
+    it('should throw error when the controller is locked', async () => {
+      await withController(async ({ controller }) => {
+        await controller.setLocked();
+
+        await expect(
+          controller.withKeyring({ type: KeyringTypes.hd }, async (keyring) =>
+            keyring.getAccounts(),
+          ),
+        ).rejects.toThrow(KeyringControllerError.ControllerLocked);
+      });
+    });
   });
 
   describe('QR keyring', () => {
@@ -2533,6 +2866,16 @@ describe('KeyringController', () => {
           expect(qrKeyring).toBeUndefined();
         });
       });
+
+      it('should throw error when the controller is locked', async () => {
+        await withController(async ({ controller }) => {
+          await controller.setLocked();
+
+          expect(() => controller.getQRKeyring()).toThrow(
+            KeyringControllerError.ControllerLocked,
+          );
+        });
+      });
     });
 
     describe('connectQRHardware', () => {
@@ -2568,6 +2911,16 @@ describe('KeyringController', () => {
           (keyring) => keyring.type === KeyringTypes.qr,
         );
         expect(qrKeyring?.accounts).toHaveLength(3);
+      });
+
+      it('should throw error when the controller is locked', async () => {
+        await withController(async ({ controller }) => {
+          await controller.setLocked();
+
+          await expect(controller.connectQRHardware(0)).rejects.toThrow(
+            KeyringControllerError.ControllerLocked,
+          );
+        });
       });
     });
 
@@ -2795,6 +3148,16 @@ describe('KeyringController', () => {
             .sign.request,
         ).toBeUndefined();
       });
+
+      it('should throw error when the controller is locked', async () => {
+        await withController(async ({ controller }) => {
+          await controller.setLocked();
+
+          await expect(controller.resetQRKeyringState()).rejects.toThrow(
+            KeyringControllerError.ControllerLocked,
+          );
+        });
+      });
     });
 
     describe('forgetQRDevice', () => {
@@ -2823,6 +3186,16 @@ describe('KeyringController', () => {
 
           expect(removedAccounts).toHaveLength(0);
           expect(remainingAccounts).toHaveLength(0);
+        });
+      });
+
+      it('should throw error when the controller is locked', async () => {
+        await withController(async ({ controller }) => {
+          await controller.setLocked();
+
+          await expect(controller.forgetQRDevice()).rejects.toThrow(
+            KeyringControllerError.ControllerLocked,
+          );
         });
       });
     });
@@ -2859,6 +3232,39 @@ describe('KeyringController', () => {
           signProcessKeyringController.state.keyrings[1].accounts,
         ).toHaveLength(1);
       });
+
+      it('should throw error when the controller is locked', async () => {
+        const serializedQRKeyring = {
+          initialized: true,
+          accounts: ['0xE410157345be56688F43FF0D9e4B2B38Ea8F7828'],
+          currentAccount: 0,
+          page: 0,
+          perPage: 5,
+          keyringAccount: 'account.standard',
+          keyringMode: 'hd',
+          name: 'Keystone',
+          version: 1,
+          xfp: '5271c071',
+          xpub: 'xpub6CNhtuXAHDs84AhZj5ALZB6ii4sP5LnDXaKDSjiy6kcBbiysq89cDrLG29poKvZtX9z4FchZKTjTyiPuDeiFMUd1H4g5zViQxt4tpkronJr',
+          hdPath: "m/44'/60'/0'",
+          childrenPath: '0/*',
+          indexes: {
+            '0xE410157345be56688F43FF0D9e4B2B38Ea8F7828': 0,
+            '0xEEACb7a5e53600c144C0b9839A834bb4b39E540c': 1,
+            '0xA116800A72e56f91cF1677D40C9984f9C9f4B2c7': 2,
+            '0x4826BadaBC9894B3513e23Be408605611b236C0f': 3,
+            '0x8a1503beb17Ef02cC4Ff288b0A73583c4ce547c7': 4,
+          },
+          paths: {},
+        };
+        await withController(async ({ controller }) => {
+          await controller.setLocked();
+
+          await expect(
+            controller.restoreQRKeyring(serializedQRKeyring),
+          ).rejects.toThrow(KeyringControllerError.ControllerLocked);
+        });
+      });
     });
 
     describe('getAccountKeyringType', () => {
@@ -2874,6 +3280,16 @@ describe('KeyringController', () => {
         expect(
           await signProcessKeyringController.getAccountKeyringType(qrAccount),
         ).toBe(KeyringTypes.qr);
+      });
+
+      it('should throw error when the controller is locked', async () => {
+        await withController(async ({ controller }) => {
+          await controller.setLocked();
+
+          await expect(controller.getAccountKeyringType('0x0')).rejects.toThrow(
+            KeyringControllerError.ControllerLocked,
+          );
+        });
       });
     });
 
@@ -2891,6 +3307,16 @@ describe('KeyringController', () => {
         await signProcessKeyringController.submitQRCryptoHDKey('anything');
         expect(submitCryptoHDKeyStub.calledWith('anything')).toBe(true);
       });
+
+      it('should throw error when the controller is locked', async () => {
+        await withController(async ({ controller }) => {
+          await controller.setLocked();
+
+          await expect(controller.submitQRCryptoHDKey('0x0')).rejects.toThrow(
+            KeyringControllerError.ControllerLocked,
+          );
+        });
+      });
     });
 
     describe('submitQRCryptoAccount', () => {
@@ -2906,6 +3332,16 @@ describe('KeyringController', () => {
         submitCryptoAccountStub.resolves();
         await signProcessKeyringController.submitQRCryptoAccount('anything');
         expect(submitCryptoAccountStub.calledWith('anything')).toBe(true);
+      });
+
+      it('should throw error when the controller is locked', async () => {
+        await withController(async ({ controller }) => {
+          await controller.setLocked();
+
+          await expect(controller.submitQRCryptoAccount('0x0')).rejects.toThrow(
+            KeyringControllerError.ControllerLocked,
+          );
+        });
       });
     });