+++ packages/transaction-controller/src/TransactionController.ts
@@ -1,7 +1,4 @@
-import { Hardfork, Common, type ChainConfig } from '@ethereumjs/common';
 import type { TypedTransaction } from '@ethereumjs/tx';
-import { TransactionFactory } from '@ethereumjs/tx';
-import { bufferToHex } from '@ethereumjs/util';
 import type { AccountsControllerGetSelectedAccountAction } from '@metamask/accounts-controller';
 import type {
   AcceptResultCallbacks,
@@ -101,6 +98,8 @@ import {
   TransactionStatus,
   SimulationErrorCode,
 } from './types';
+import type { KeyringControllerSignAuthorization } from './utils/eip7702';
+import { signAuthorizationList } from './utils/eip7702';
 import { validateConfirmedExternalTransaction } from './utils/external-transactions';
 import { addGasBuffer, estimateGas, updateGas } from './utils/gas';
 import { updateGasFees } from './utils/gas-fees';
@@ -117,6 +116,7 @@ import {
   getAndFormatTransactionsForNonceTracker,
   getNextNonce,
 } from './utils/nonce';
+import { prepareTransaction, serializeTransaction } from './utils/prepare';
 import { getTransactionParamsWithIncreasedGasFee } from './utils/retry';
 import { getSimulationData } from './utils/simulation';
 import {
@@ -161,7 +161,6 @@ const metadata = {
   },
 };
 
-export const HARDFORK = Hardfork.London;
 const SUBMIT_HISTORY_LIMIT = 100;
 
 /**
@@ -343,10 +342,11 @@ const controllerName = 'TransactionController';
  * The external actions available to the {@link TransactionController}.
  */
 export type AllowedActions =
+  | AccountsControllerGetSelectedAccountAction
   | AddApprovalRequest
+  | KeyringControllerSignAuthorization
   | NetworkControllerFindNetworkClientIdByChainIdAction
-  | NetworkControllerGetNetworkClientByIdAction
-  | AccountsControllerGetSelectedAccountAction;
+  | NetworkControllerGetNetworkClientByIdAction;
 
 /**
  * The external events available to the {@link TransactionController}.
@@ -583,7 +583,7 @@ export class TransactionController extends BaseController<
   TransactionControllerState,
   TransactionControllerMessenger
 > {
-  #internalEvents = new EventEmitter();
+  readonly #internalEvents = new EventEmitter();
 
   private readonly isHistoryDisabled: boolean;
 
@@ -593,7 +593,7 @@ export class TransactionController extends BaseController<
 
   private readonly approvingTransactionIds: Set<string> = new Set();
 
-  #methodDataHelper: MethodDataHelper;
+  readonly #methodDataHelper: MethodDataHelper;
 
   private readonly mutex = new Mutex();
 
@@ -622,9 +622,9 @@ export class TransactionController extends BaseController<
     chainId?: string,
   ) => NonceTrackerTransaction[];
 
-  #incomingTransactionChainIds: Set<Hex> = new Set();
+  readonly #incomingTransactionChainIds: Set<Hex> = new Set();
 
-  #incomingTransactionHelper: IncomingTransactionHelper;
+  readonly #incomingTransactionHelper: IncomingTransactionHelper;
 
   private readonly layer1GasFeeFlows: Layer1GasFeeFlow[];
 
@@ -638,15 +638,15 @@ export class TransactionController extends BaseController<
 
   private readonly signAbortCallbacks: Map<string, () => void> = new Map();
 
-  #trace: TraceCallback;
+  readonly #trace: TraceCallback;
 
-  #transactionHistoryLimit: number;
+  readonly #transactionHistoryLimit: number;
 
-  #isFirstTimeInteractionEnabled: () => boolean;
+  readonly #isFirstTimeInteractionEnabled: () => boolean;
 
-  #isSimulationEnabled: () => boolean;
+  readonly #isSimulationEnabled: () => boolean;
 
-  #testGasFeeFlows: boolean;
+  readonly #testGasFeeFlows: boolean;
 
   private readonly afterSign: (
     transactionMeta: TransactionMeta,
@@ -721,7 +721,7 @@ export class TransactionController extends BaseController<
     );
   }
 
-  #multichainTrackingHelper: MultichainTrackingHelper;
+  readonly #multichainTrackingHelper: MultichainTrackingHelper;
 
   /**
    * Method used to sign transactions
@@ -1039,19 +1039,25 @@ export class TransactionController extends BaseController<
       );
     }
 
+    const permittedAddresses =
+      origin === undefined
+        ? undefined
+        : await this.getPermittedAccounts?.(origin);
+
+    const selectedAddress = this.#getSelectedAccount().address;
+
+    await validateTransactionOrigin({
+      from: txParams.from,
+      origin,
+      permittedAddresses,
+      selectedAddress,
+      txParams,
+    });
+
     const isEIP1559Compatible =
       await this.getEIP1559Compatibility(networkClientId);
 
     validateTxParams(txParams, isEIP1559Compatible);
-
-    if (origin && this.getPermittedAccounts) {
-      await validateTransactionOrigin(
-        await this.getPermittedAccounts(origin),
-        this.#getSelectedAccount().address,
-        txParams.from,
-        origin,
-      );
-    }
 
     const dappSuggestedGasFees = this.generateDappSuggestedGasFees(
       txParams,
@@ -1331,7 +1337,7 @@ export class TransactionController extends BaseController<
 
     prepareTransactionParams?.(newTxParams);
 
-    const unsignedEthTx = this.prepareUnsignedEthTx(
+    const unsignedEthTx = prepareTransaction(
       transactionMeta.chainId,
       newTxParams,
     );
@@ -1346,7 +1352,7 @@ export class TransactionController extends BaseController<
       signedTx,
     );
 
-    const rawTx = bufferToHex(signedTx.serialize());
+    const rawTx = serializeTransaction(signedTx);
     const newFee = newTxParams.maxFeePerGas ?? newTxParams.gasPrice;
 
     const oldFee = newTxParams.maxFeePerGas
@@ -1928,18 +1934,14 @@ export class TransactionController extends BaseController<
 
     const initialTx = listOfTxParams[0];
     const { chainId } = initialTx;
-    const common = this.getCommonConfiguration(chainId);
     const networkClientId = this.#getNetworkClientId({ chainId });
-
-    const initialTxAsEthTx = TransactionFactory.fromTxData(initialTx, {
-      common,
-    });
-
-    const initialTxAsSerializedHex = bufferToHex(initialTxAsEthTx.serialize());
+    const initialTxAsEthTx = prepareTransaction(chainId, initialTx);
+    const initialTxAsSerializedHex = serializeTransaction(initialTxAsEthTx);
 
     if (this.approvingTransactionIds.has(initialTxAsSerializedHex)) {
       return '';
     }
+
     this.approvingTransactionIds.add(initialTxAsSerializedHex);
 
     let rawTransactions, nonceLock;
@@ -2248,14 +2250,15 @@ export class TransactionController extends BaseController<
     };
 
     const { from } = updatedTransactionParams;
-    const common = this.getCommonConfiguration(chainId);
-    const unsignedTransaction = TransactionFactory.fromTxData(
-      updatedTransactionParams,
-      { common },
-    );
-    const signedTransaction = await this.sign(unsignedTransaction, from);
 
-    const rawTransaction = bufferToHex(signedTransaction.serialize());
+    const unsignedTransaction = prepareTransaction(
+      chainId,
+      updatedTransactionParams,
+    );
+
+    const signedTransaction = await this.sign(unsignedTransaction, from);
+    const rawTransaction = serializeTransaction(signedTransaction);
+
     return rawTransaction;
   }
 
@@ -2274,6 +2277,7 @@ export class TransactionController extends BaseController<
   /**
    * Stop the signing process for a specific transaction.
    * Throws an error causing the transaction status to be set to failed.
+   *
    * @param transactionId - The ID of the transaction to stop signing.
    */
   abortTransactionSigning(transactionId: string) {
@@ -2562,18 +2566,17 @@ export class TransactionController extends BaseController<
           note: 'TransactionController#approveTransaction - Transaction approved',
         },
         (draftTxMeta) => {
-          const { txParams, chainId } = draftTxMeta;
+          const { chainId, txParams } = draftTxMeta;
+          const { gas, type } = txParams;
 
           draftTxMeta.status = TransactionStatus.approved;
-          draftTxMeta.txParams = {
-            ...txParams,
-            nonce,
-            chainId,
-            gasLimit: txParams.gas,
-            ...(isEIP1559Transaction(txParams) && {
-              type: TransactionEnvelopeType.feeMarket,
-            }),
-          };
+          draftTxMeta.txParams.chainId = chainId;
+          draftTxMeta.txParams.gasLimit = gas;
+          draftTxMeta.txParams.nonce = nonce;
+
+          if (!type && isEIP1559Transaction(txParams)) {
+            draftTxMeta.txParams.type = TransactionEnvelopeType.feeMarket;
+          }
         },
       );
 
@@ -2718,8 +2721,6 @@ export class TransactionController extends BaseController<
       updatedTransactionMeta,
     );
     this.#internalEvents.emit(
-      // TODO: Either fix this lint violation or explain why it's necessary to ignore.
-      // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
       `${transactionMeta.id}:finished`,
       updatedTransactionMeta,
     );
@@ -2755,8 +2756,6 @@ export class TransactionController extends BaseController<
         const { chainId, status, txParams, time } = tx;
 
         if (txParams) {
-          // TODO: Either fix this lint violation or explain why it's necessary to ignore.
-          // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
           const key = `${String(txParams.nonce)}-${convertHexToDecimal(
             chainId,
           )}-${new Date(time).toDateString()}`;
@@ -2920,35 +2919,6 @@ export class TransactionController extends BaseController<
       chainId,
       networkClientId,
     }).provider;
-  }
-
-  private prepareUnsignedEthTx(
-    chainId: Hex,
-    txParams: TransactionParams,
-  ): TypedTransaction {
-    return TransactionFactory.fromTxData(txParams, {
-      freeze: false,
-      common: this.getCommonConfiguration(chainId),
-    });
-  }
-
-  /**
-   * `@ethereumjs/tx` uses `@ethereumjs/common` as a configuration tool for
-   * specifying which chain, network, hardfork and EIPs to support for
-   * a transaction. By referencing this configuration, and analyzing the fields
-   * specified in txParams, @ethereumjs/tx is able to determine which EIP-2718
-   * transaction type to use.
-   *
-   * @param chainId - The chainId to use for the configuration.
-   * @returns common configuration object
-   */
-  private getCommonConfiguration(chainId: Hex): Common {
-    const customChainParams: Partial<ChainConfig> = {
-      chainId: parseInt(chainId, 16),
-      defaultHardfork: HARDFORK,
-    };
-
-    return Common.custom(customChainParams);
   }
 
   private onIncomingTransactions(transactions: TransactionMeta[]) {
@@ -3206,9 +3176,18 @@ export class TransactionController extends BaseController<
   ): Promise<string | undefined> {
     log('Signing transaction', txParams);
 
-    const unsignedEthTx = this.prepareUnsignedEthTx(
+    const { authorizationList, from } = txParams;
+    const finalTxParams = { ...txParams };
+
+    finalTxParams.authorizationList = await signAuthorizationList({
+      authorizationList,
+      messenger: this.messagingSystem,
+      transactionMeta,
+    });
+
+    const unsignedEthTx = prepareTransaction(
       transactionMeta.chainId,
-      txParams,
+      finalTxParams,
     );
 
     this.approvingTransactionIds.add(transactionMeta.id);
@@ -3216,7 +3195,7 @@ export class TransactionController extends BaseController<
     const signedTx = await new Promise<TypedTransaction>((resolve, reject) => {
       this.sign?.(
         unsignedEthTx,
-        txParams.from,
+        from,
         ...this.getAdditionalSignArguments(transactionMeta),
       ).then(resolve, reject);
 
@@ -3256,7 +3235,7 @@ export class TransactionController extends BaseController<
 
     this.onTransactionStatusChange(transactionMetaWithRsv);
 
-    const rawTx = bufferToHex(signedTx.serialize());
+    const rawTx = serializeTransaction(signedTx);
 
     const transactionMetaWithRawTx = merge({}, transactionMetaWithRsv, {
       rawTx,
@@ -3405,21 +3384,12 @@ export class TransactionController extends BaseController<
     return pendingTransactionTracker;
   }
 
-  #checkForPendingTransactionAndStartPolling = () => {
+  readonly #checkForPendingTransactionAndStartPolling = () => {
     this.#multichainTrackingHelper.checkForPendingTransactionAndStartPolling();
   };
 
   #stopAllTracking() {
     this.#multichainTrackingHelper.stopAllTracking();
-  }
-
-  #removeIncomingTransactionHelperListeners(
-    incomingTransactionHelper: IncomingTransactionHelper,
-  ) {
-    incomingTransactionHelper.hub.removeAllListeners('transactions');
-    incomingTransactionHelper.hub.removeAllListeners(
-      'updated-last-fetched-timestamp',
-    );
   }
 
   #addIncomingTransactionHelperListeners(